requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
click==8.3.1
exceptiongroup==1.3.1
fastapi==0.123.10
greenlet==3.3.0
h11==0.16.0
idna==3.11
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
SQLAlchemy==2.0.44
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
